homework_1/api/requirements.txt
fastapi[all]==0.104.1
names==0.3.0
python-dotenv==1.0.0
randomtimestamp==2.2
